Changes for 01.01.2023 Added BoomerangView to separate view
