+++ AnimalCard.xcodeproj/xcuserdata/vardant.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "B0CAB21A-FF04-470F-9193-B0CDD8080031"
+   type = "1"
+   version = "2.0">
+</Bucket>
